fix seeds

+++ database/migrations/2019_09_07_172853_create_table_products.php
@@ -17,7 +17,7 @@ class CreateTableProducts extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->float('price');
-            $table->float('quantidade');
+            $table->integer('quantidade');
             $table->text('description');
             $table->timestamps();
         });
